Added 'missed things' to theme 'minima' (sic!) to have layout and CSS.
Long story short - according to this issue:
https://github.com/jekyll/minima/issues/38
you HAVE to copy data from default theme to have a theme.
Lovely solution. In other words: set theme in 'Gemfile' is not enough
to have that theme. And the funniest things is that, if you rebuild
it locally everything look OK, only on github you don't have CSS.
'